+++ src/commonMain/kotlin/csense/kotlin/tests/assertions/CharSequence.kt
@@ -3,7 +3,14 @@
 package csense.kotlin.tests.assertions
 
 
-public inline fun CharSequence.assert(other: CharSequence, ignoreCase: Boolean = false) {
+public inline fun CharSequence?.assert(other: CharSequence?, ignoreCase: Boolean = false) {
+    if (this == null || other == null) {
+        if (this == other) {
+            return
+        } else {
+            failTest("Expected $other but got $this")
+        }
+    }
     this.length.assert(
         other.length, message = "expected length to be equal for both " +
                 "actual \"$this\"(length = ${this.length}) and the " +