+++ src/commonMain/kotlin/csense/kotlin/tests/assertions/MapEntry.kt
@@ -3,10 +3,11 @@
 package csense.kotlin.tests.assertions
 
 
-public inline fun <@kotlin.internal.OnlyInputTypes Key, @kotlin.internal.OnlyInputTypes Value> Map.Entry<Key, Value>.assert(
+public inline fun <@kotlin.internal.OnlyInputTypes Key, @kotlin.internal.OnlyInputTypes Value> Map.Entry<Key, Value>?.assert(
     key: Key,
     value: Value
 ) where Key : Comparable<Key>, Value : Comparable<Value> {
+    assertNotNull()
     this.key.assert(key)
     this.value.assert(value)
 }